Extract unbounded reader
